-Criado Classe utilitaria para centralização de funções em comum -Alterado caminho das funções nos arquivos da recepção
