+++ Gabinete/src/java/br/edu/ifrs/modelo/bean/Recepcao.java
@@ -130,30 +130,4 @@ public class Recepcao
     public void setCpf_servidor(String cpf_servidor) {
         this.cpf_servidor = cpf_servidor;
     }
-    
-    public static Calendar formataStringToCalendar(String data) throws Exception 
-    { 
-        SimpleDateFormat formatoData = new SimpleDateFormat("dd/mm/yyyy");
-        Calendar c = Calendar.getInstance();
-        if (data == null || data.equals(""))
-            return null;
-        try 
-        {
-            c.setTime(formatoData.parse(data));
-        } catch (ParseException e) {            
-            throw e;
-        }
-        return c;
-    }    
-    
-    public static String formataCalendarToString(Calendar data) throws Exception 
-    { 
-        SimpleDateFormat formatoData = new SimpleDateFormat("yyyy-mm-dd");
-        Calendar c = new GregorianCalendar();
-        String retorno="";
-        if (data == null || data.equals(""))
-            return null;
-        retorno = formatoData.format(c.getTime());
-        return retorno;
-    }    
 }
